+++ opal/opal-jquery/http.rb
@@ -6,6 +6,15 @@ require 'opal-jquery/constants'
 class HTTP
   `var $ = #{JQUERY_SELECTOR.to_n}` # cache $ for SPEED
 
+  def self.setup
+    Hash.new(`$.ajaxSetup()`)
+  end
+
+  def self.setup= settings
+    `$.ajaxSetup(#{settings.to_n})`
+  end
+
+
   attr_reader :body, :error_message, :method, :status_code, :url, :xhr
 
   def self.get(url, opts={}, &block)